migrate unit_test_log to pytest
